Add arbitrary manpage/infopage/source compression support (#142).
Based on patch from pkarlsen at mandriva.com.

The UseBzip2, UseLzma, and UseXz config options have been replaced by
the new CompressExtension option (which takes a filename extension
without the leading dot), the *page-not-compressed-with-* message ids
have been replaced by *page-not-compressed, and the
source-or-patch-not-*zipped message ids have been replaced by
source-or-patch-not-compressed.


git-svn-id: svn+ssh://rpmlint.zarb.org/home/projects/rpmlint/svn/trunk@1707 9bc8b190-ac0f-0410-8968-dc7d1f502856