warn when a pem certificate is found in a package.
after the compromise of various CA in the last years, it would
be better to not have X copy of the CA certificate in each package.

Not to mention the issue of having some packages shipping a default
private key with the certificate for a perfect out of box experience, that
could lead to security issue. So the check look for certificate, but
verify that the file is not not marked as %ghost ).

idea from kiilerix on irc