add /usr/lib/systemd to DEFAULT_HARDCODED_LIB_PATH_EXCEPTIONS (#93)
Spec files that need to install files into, e.g.,
/usr/lib/systemd/system-generators will generate a hardcoded-library-path
error.  This commit adds /usr/lib/systemd to the list of exceptions for this
check. In modern systems hardcoding these paths can be avoided by using macros,
but all the macros are not available in older systems
(e.g. %_systemdgeneratordir on CentOS/RHEL 7).