get_default_valid_rpmgroups: Don't fail if rpm is partially installed
A catch-all exception was added to handle the case where the rpm package
was not installed dec4821 ("get_default_valid_rpmgroups: Don't fail if
the rpm package is not installed", 2015-07-24).  That catch-all was
later tightened in c1945e3 ("Avoid catch-all except statements",
2017-10-24).

The rpm package can also trigger a TypeError exception when installed
without docs or all languages, as is commonly done in container images.
Catch TypeError exception as well.