Treat Sphinx inventory files as binary
Sphinx, a tool commonly used to produce documentation for Python
projects, produces inventory files with extension .inv that contain a
plaintext header followed by raw zlib data. When the zlib data is short
enough, this fools the heuristic used to determine if a file is text or
binary. Since these files are often included in documentation packages,
this can produce false positives for the wrong-file-end-of-line-encoding
check and the subsequent file-not-utf8 check.