Say goodbye to manpage and bash_completion
Sadly both of those are really obsolete and not matching the behaviour
that is currently present in the binaries.

The manpage content is now in argparse data when one runs --help it
provides equal information that was stored in the mans.

If needed one can use help2man to generate new ones during the
package build