Remove LSBCheck
The purpose of this check was to validate name, version and release
according to LSB. It was added 19 years ago and it hasn't been updated
since that time.

It's not clear where the stated rules for name/version/release comes
from. Moreover, there is a mismatch between regexps used and
a description in toml file so it's not clear what it's the
authoritative source and we can't validate it.