Separate TmpFilesCheck to functions and add a test
Separate the check() method to multiple methods and clean up the code.
Also add a test (systemd-tmpfiles-1.0-0.x86_64.rpm) for an additional
testing.

Add a method that checks that "systemd-tmpfiles --create" call
is not present in the %pre section as it should be in the %post.

Remove "now check remaining ghost files that are not already" part
as the same code is already present in PostCheck.py and it's not related
to tmpfiles anyway. This means that "ghost-files-without-postin" and
"postin-without-ghost-file-creation" checks were removed from
TmpFilesCheck.py.

Rename postin-without-tmpfile-creation to post-without-tmpfile-creation
as we want to address %post section (there is no %postin).