Remove NamingPolicyCheck
After reviewing the check the only really valid naming schemes
applicable are for ocaml and apache modules. The others are actively
misleading to namingspace on Fedora and openSUSE.

The ruby/php/python/... well most bindings are sometimes versioned
if multiversioning is needed and sometimes not, so these policies
are actively misleading.

The better soltuion is to use reviewer to properly request what
naming should be used for the package.