Move the rpmlint CLI logic into its own module file
Having the CLI processing logic in the init module file made it
get evaluated on every single rpmlint import. This is completely
unnecessary and causes an import loop in setuptools, introducing
dependencies where there should not be when trying to "build" the
rpmlint code with setuptools.

This should also speed up imports across the board in rpmlint, since
this code only needs to be evaluated once.