Support overriding error descriptions in config
Now we will allow adding new descriptions strings for the rpmlint
errors in configuration files. The new error descriptions can be
defined in TOML syntax as [Descriptions] table with key/value pairs,
where key is the rpmlint error name and value is the new text for
that error.

Example:

[Descriptions]
no-binary = """
A new text for no-binary error.
"""