FakePkg: fix paths in add_file_with_content()
The path is wrongly assigned to the FakePkg object, not to the PkgFile
object. This causes actual file access to fail.

Along with this also use filenames derived from the actual file path.
This allows to inspect the temporary directory and get some sensible
filenames instead of just temporary files. The TemporaryDirectory will
be recursively deleted at the end anyway.