FileDigestCheck: remove -digest- from error names where unnecessary
The error names "<group>-file-digest-unauthorized" and
"<group>-file-digest-ghost" are a bit confusing, because paths can also
be whitelisted by wildcards without adding any digests. Also this
exposes an implementation detail, namely that a digest _could_ be used
in the end for whitelisting. Therefore keep the digest name component
only for the case when a file digest is actually mismatching, otherwise
remove it.