Rework the lib_regex pattern
The pattern is meant to match with filenames that (look like they)
constitute a _versioned_ library, for which there are package naming
policies (openSUSE), and/or required ldconfig calls. Anything not
matched by this regex is going to be a non-versioned library or not a
library at all.

The pattern so far failed to match on "libSPIRV-tools-opt.so.suse22",
which is a legit versioned library in openSUSE. In coming up with a
new regex, the following considerations came to mind:

1. Library filenames should really just consist of few characters,
hence we strictly stick to chars like [\w-.] (If your library
filename has Unicode or square brackets, you should rethink your life
choices.)

2. Central name ought not to start/end with dash or underscore, hence
we use [A-Za-z0-9] in some places instead of \w (\w includes _).

2a. lib4ti2.so is a thing, and so digits as first character need to
be supported :-/

(3. Vision statement: ELF files should have a field to distinguish
their intended purpose (unspecified; intended for ld -lxxx and/or
else, intended only for dlopen, intended only for LD_PRELOAD, ...),
which would help in the determination whether to apply policies.)

4. Versioning exists after ".so.": great

4a. Dashes in the SOVER portion are very unusual, which is why the
(new) pattern does not allow for it. One (desired) side effect of
this is that gdb prettyprinters (libc.so.6-gdb.py) get treated as
not-versioned.

5. Versioning only before ".so": Because letters can constitute
versioning just as much as digits, "libdb-fourthgen.so" can be legit
versioning (in lieu of a more "normal" expectation like
"libdb-4.so"). Equally, "libdb-4.so" could be a plugin whose name
just happens to have digits. It is impossible to tell from the name
alone whether something is versioned or unversioned.

So then, a heuristic it shall be.

* libauth-samba4.so: unversioned (though it has ELF symbol versions, the filename
  is unversioned)
* /usr/lib/gconv/ISO_6937.so: unversion (plugin to iconv)
* libdb_4.so: unversioned (_\d is not a common sight at all)
* libdb-4.so: versioned (-\d is definitely a common sight, e.g.
  libbfd, libglib, etc.)