Prevent duplicate files in lint input
If the same file is passed in more than once as lint input,
certain checks misbehave.
At least error multiple-specfiles can be triggered
simply by passing in the same srpm twice.
Due to globbing, duplicates may be unknowingly passed in.
In order to avoid needless errors when that happens,
the rpmfile list in lint input is deduplicated
by collecting them in a set instead of a list.