FileDigestCheck: implement support for file digest filtering
We want to become more robust against trivial changes in whitelisted
files. Currently we are not using file digests for things like D-Bus
service XML files, because a lot of non-functional changes might be
disrupting our processes.

With these filter we will be able to avoid this and become aware of more
sensitive upstream changes in the future.

This commit only contains the infrastucture, no production whitelists
use the feature yet, they will rely on the DefaultDigester until we
decide how to continue from here.