SystemdTmpfilesCheck: new check to restrict systemd-tmpfiles configuration
This introduces a new check that restricts what kind of systemd-tmpfiles
configuration file entries can be shipped by a package. This supplements
the existing security related checkers like FileDigestCheck. Since this
check requires parsing of the custom systemd-tmpfiles format the generic
existing checkers cannot be reused for this.

This change does not yet raise badness in the strict configuration,
because the check first needs some realife testing and the affected
packages first need to be reviewed and whitelisted before actually
generating badness.