checkbashism: fix stuck when run with Emacs
Apparently, when shell=True is used for subprocess::check_output, one
is supposed to not using sequence as first argument:

```
The shell argument (which defaults to False) specifies whether to use the shell as the program to execute.
If shell is True, it is recommended to pass args as a string rather than as a sequence.
```

Fixes: #886