configs/Fedora: sync user/groups with setup-2.14.1
This removes 3 users/groups:

    * #systemd-journal-gateway
    * ergo
    * katello

The `#systemd-journal-gateway` entry was due to not excluding comment
lines in the current uidgid file.  It was removed upstream in a2b68d4
(remove uidgid reservation for systemd-journal-gateway (#1174304),
2014-12-18)¹.

The `ergo` entry was added in ab36f34e (Add ergo user and group,
2022-04-13)².  However, the Fedora ergo⁴ package is a "program for
large-scale self-consistent field calculations."  The ergo⁵ package
which was referenced in #842 is an IRC daemon written in Go, which is
not in Fedora.

The `katello` entry was removed upstream in 6563bed (Stop creating
katello user, 2021-06-16)³.

This was done with tools/generate-fedora-users-groups.py:

    tools/generate-fedora-users-groups.py >configs/Fedora/users-groups.toml

¹ https://pagure.io/setup/c/a2b68d4
² https://github.com/rpm-software-management/rpmlint/commit/ab36f34e
  https://github.com/rpm-software-management/rpmlint/pull/842
³ https://pagure.io/setup/c/6563bed
⁴ https://src.fedoraproject.org/rpms/ergo
⁵ https://github.com/ergochat/ergo