rpmdiff: improve help for multiple values with -i/-e options
When using multiple values with the -i or -e options, rpmlint-2.x
is incompatible with earlier releases.  Moreover, the behavior is not
immediately obvious to users¹.

To help alleviate this slightly, add an epilog to the parser which
explains how to pass multiple options to -i/-e options using either '--'
to separate the values from the following positional arguments or lists
the options _after_ the positional arguments.

This isn't really a fix for the poor behavior we get from argparse
relative to what older releases had with getopt, but maybe it's enough
until folks get used to how the options must be handled in 2.x releases.

This partially addresses #940.

¹ https://bugzilla.redhat.com/1969534