openSUSE: new whitelisting restriction for sysctl drop-in files (bsc#1174722)
Based on the file digest whitelisting check this introduces a new
whitelisting restriction for installation of sysctl.d drop-in
configuration files. The files whitelisted here have been reviewed.

Content based whitelistings seem feasible since there is only a small
number of packages that ship files for this and the ones that do only
rarely change in content.