FileDigestCheck: support exceptions to symlink restrictions
For some checks it would be unfortunate to generally have to deal with
the possibility of symlinks just because a single corner case exists.
The new sysctl restriction is such a case, where systemd ships a
backward compatibility symlink.

Since this symlink cannot be verified anyway (the target file is part of
another package) it seems cleaner to simply ignore this one occurence of
a symlink but otherwise still reject them.

For this use a mechanism similar to the one already present for %ghost
files.