dbus-services: remove no longer packaged mumble-server (bsc#1209338)
This package was dropped for a while from Factory and the re-addition
triggered a review due to changed paths from /etc to /usr. It turned out
that the system wide D-Bus service is not really needed at all, though,
so it got dropped.

For some reason there are duplicate entries for this package, probably a
bug from the porting from rpmlint1 to rpmlint2.