SUIDPermissionsCheck: fix broken warnings and harmonize code
At some point the conditions for this check changed, causing it to
currently always trigger 'permissions-dir-without-slash' (false
positives).

The reason for this is that the keys in `self.perms` are always
normalized now via the `PermissionsParser` helper. This means there
won't ever a trailing slash be found in the keys of `self.perms`.
Instead the trailing slash is conserved in the entry object found in the
values of `self.perms`.

To fix this check the `entry.path` for existence of the trailing slash,
not the keys. Also harmonize the rest of the code:

- _exists_permissions_entry() is no longer necessary, since the keys are
  always normalized.
- check for both 'permissions-dir-without-slash' and
  'permissions-file-as-dir' within `_verify_entry()`.