lint: Reset all checks for each package
Some check classes keep state during checking and this cause problems
when checking several files with the same call.

This patch resets all the checks for each package to check. This is a
quick solution, maybe the best way to handle this is to reset the state
for each check in with a common method in the AbstractCheck, a common
way to handle and reset the state to avoid this problem in the future.

Fix https://github.com/rpm-software-management/rpmlint/issues/1161