FilesCheck: Fix zero permission check with folder
The zero permission check tries to read tmpfiles.d entries, but if a package
has a folder with that name it crashes trying to open the folder as a file to
read.

This patch fix the issue and check if the file is a directory before trying to
open it for reading.

Fix https://github.com/rpm-software-management/rpmlint/issues/1257