LogrotateCheck: consider configuration files in /usr/etc; add badness
I noticed a mismatch between rpmlint1 in SUSE:SLE-15-SP7 and rpmlint2 in
Factory regarding the "exim" package. It ships a logrotate configuration
file to operate as root in a log directory owned by non-root.

In SLE-15-SP7, where the configuration file is installed into
/etc/logrotate.d, this is an error diagnostic. In Factory, however, the
configuration file is installed into /usr/etc, which is not considered
by the LogrotateCheck at all. Fix that.

Since the user-writable-log-dir is a security relevant issue (logrotate
will operate as root in a non-root owned directory), we should
add badness for this diagnostic, as well. Since this might hit many
packages let's start with low badness. Then we can inspect which
packages are affected to give them some heads-up. When everything is
clear we can raise the badness to 1,000 as usual for security relevant
diagnostics.