Run only once the validate_filters per call
When the cli tool is called with "-i" option in combination with a list of
spec/rpm files it runs the validate_installed_packages and also the
validate_files. These methods check the list of packages to lint to just run
post check function and validate used filters in rpmlintrc.

This patch makes sure to do not call the post check and validation twice,
checking if there are also "files" during the validate_installed_packages. That
way the is_last will be only true in the latter call inside validate_files.

Fix https://github.com/rpm-software-management/rpmlint/issues/1368