opensuse.toml: block filters for sudoers-, sysctl- whitelisting errors
When these new restrictions have been added in commit 171c7ad1b7c925
and commit e9d5a19fe90b, we've overlooked that the diagnostics are not
in the BlockedFilters list. This means packagers can currently reduce
the badness of these warnings via their *.rpmlintrc files.

I checked the current list of rpmlint diagnostics on
rpmlint.opensuse.org and luckily nobody seems to make use of this
loophole at the moment.

Add the missing diagnostics to the list to sync everything now.